Refactor reports API and add unit tests (#770)
* Refactor web.go into api/schema.go

* Copy init() fn to schema.go

* Refactor reports API and add unit tests

* Add missing types.go

* revert changes on rules_test
